cdss/requirements.txt
funasr==1.2.6
paddlepaddle==3.0.0b1
paddleocr==2.9.1
fastapi==0.115.6
openai==1.59.6
argparse==1.4.0
torch==2.5.1
torchaudio==2.5.1
python-multipart==0.0.20
pydantic-settings==2.2.1
pyyaml==6.0.1
dotenv==0.9.9
uvicorn==0.34.0